Added guards to the program-specific functions
This means that the functions/aliases that are build around a certain
command (eg git) will not be available if said command is not available.

+++ zsh/rc/functions/git.zsh
@@ -1,3 +1,7 @@
+if ! command -v git &> /dev/null; then
+	return
+fi
+
 # Some shortcuts.
 alias gits='rm -f \\ && git status'
